+++ packages/playground/vue/setup-import-template/SetupImportTemplate.vue
@@ -1,0 +1,5 @@
+<template src="./template.html"></template>
+<script setup>
+let count = $ref(0)
+const inc = () => count++
+</script>