+++ packages/vite/src/node/config.ts
@@ -7,6 +7,11 @@ import {
   resolveServerOptions,
   ServerOptions
 } from './server'
+import {
+  ResolvedPreviewOptions,
+  resolvePreviewOptions,
+  PreviewOptions
+} from './preview'
 import { CSSOptions } from './plugins/css'
 import {
   arraify,
@@ -141,6 +146,10 @@ export interface UserConfig {
    */
   build?: BuildOptions
   /**
+   * Preview specific options, e.g. host, port, https...
+   */
+  preview?: PreviewOptions
+  /**
    * Dep optimization options
    */
   optimizeDeps?: DepOptimizationOptions
@@ -225,6 +234,7 @@ export type ResolvedConfig = Readonly<
     plugins: readonly Plugin[]
     server: ResolvedServerOptions
     build: ResolvedBuildOptions
+    preview: ResolvedPreviewOptions
     assetsInclude: (file: string) => boolean
     logger: Logger
     createResolver: (options?: Partial<InternalResolveOptions>) => ResolveFn
@@ -354,7 +364,7 @@ export async function resolveConfig(
 
   // resolve public base url
   const BASE_URL = resolveBaseUrl(config.base, command === 'build', logger)
-  const resolvedBuildOptions = resolveBuildOptions(config.build)
+  const resolvedBuildOptions = resolveBuildOptions(resolvedRoot, config.build)
 
   // resolve cache directory
   const pkgPath = lookupFile(
@@ -405,7 +415,7 @@ export async function resolveConfig(
             ]
           }))
       }
-      return (await container.resolveId(id, importer, undefined, ssr))?.id
+      return (await container.resolveId(id, importer, { ssr }))?.id
     }
   }
 
@@ -417,6 +427,8 @@ export async function resolveConfig(
           typeof publicDir === 'string' ? publicDir : 'public'
         )
       : ''
+
+  const server = resolveServerOptions(resolvedRoot, config.server)
 
   const resolved: ResolvedConfig = {
     ...config,
@@ -432,8 +444,9 @@ export async function resolveConfig(
     mode,
     isProduction,
     plugins: userPlugins,
-    server: resolveServerOptions(resolvedRoot, config.server),
+    server,
     build: resolvedBuildOptions,
+    preview: resolvePreviewOptions(config.preview, server),
     env: {
       ...userEnv,
       BASE_URL,
@@ -860,8 +873,7 @@ export async function loadConfigFromFile(
     if (!userConfig) {
       // 2. if we reach here, the file is ts or using es import syntax, or
       // the user has type: "module" in their package.json (#917)
-      // transpile es import syntax to require syntax using rollup.
-      // lazy require rollup (it's actually in dependencies)
+      // transpile es import syntax to require syntax using esbuild.
       const bundled = await bundleConfigFile(resolvedPath)
       dependencies = bundled.dependencies
       userConfig = await loadConfigFromBundledFile(resolvedPath, bundled.code)