+++ packages/vite/src/node/plugins/resolve.ts
@@ -13,7 +13,6 @@ import {
   isBuiltin,
   bareImportRE,
   createDebugger,
-  deepImportRE,
   injectQuery,
   isExternalUrl,
   isObject,
@@ -24,7 +23,11 @@ import {
   isDataUrl,
   cleanUrl,
   slash,
-  nestedResolveFrom
+  nestedResolveFrom,
+  isFileReadable,
+  isTsRequest,
+  isPossibleTsOutput,
+  getTsSrcPath
 } from '../utils'
 import { ViteDevServer, SSROptions } from '..'
 import { createFilter } from '@rollup/pluginutils'
@@ -63,7 +66,14 @@ export interface InternalResolveOptions extends ResolveOptions {
   tryPrefix?: string
   skipPackageJson?: boolean
   preferRelative?: boolean
+  preserveSymlinks?: boolean
   isRequire?: boolean
+  // #3040
+  // when the importer is a ts module,
+  // if the specifier requests a non-existent `.js/jsx/mjs/cjs` file,
+  // should also try import from `.ts/tsx/mts/cts` source file as fallback.
+  isFromTsImporter?: boolean
+  tryEsmOnly?: boolean
 }
 
 export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
@@ -74,10 +84,6 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
     ssrConfig,
     preferRelative = false
   } = baseOptions
-  const requireOptions: InternalResolveOptions = {
-    ...baseOptions,
-    isRequire: true
-  }
   let server: ViteDevServer | undefined
 
   const { target: ssrTarget, noExternal: ssrNoExternal } = ssrConfig ?? {}
@@ -89,7 +95,8 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
       server = _server
     },
 
-    resolveId(id, importer, resolveOpts, ssr) {
+    resolveId(id, importer, resolveOpts) {
+      const ssr = resolveOpts?.ssr === true
       if (id.startsWith(browserExternalId)) {
         return id
       }
@@ -102,22 +109,22 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
       const targetWeb = !ssr || ssrTarget === 'webworker'
 
       // this is passed by @rollup/plugin-commonjs
-      const isRequire =
-        resolveOpts &&
-        resolveOpts.custom &&
-        resolveOpts.custom['node-resolve'] &&
-        resolveOpts.custom['node-resolve'].isRequire
+      const isRequire: boolean =
+        resolveOpts?.custom?.['node-resolve']?.isRequire ?? false
 
-      const options = isRequire ? requireOptions : baseOptions
+      const options: InternalResolveOptions = {
+        ...baseOptions,
 
-      const preserveSymlinks = !!server?.config.resolve.preserveSymlinks
+        isRequire,
+        isFromTsImporter: isTsRequest(importer ?? '')
+      }
 
       let res: string | PartialResolvedId | undefined
 
       // explicit fs paths that starts with /@fs/*
       if (asSrc && id.startsWith(FS_PREFIX)) {
         const fsPath = fsPathFromId(id)
-        res = tryFsResolve(fsPath, options, preserveSymlinks)
+        res = tryFsResolve(fsPath, options)
         isDebug && debug(`[@fs] ${chalk.cyan(id)} -> ${chalk.dim(res)}`)
         // always return here even if res doesn't exist since /@fs/ is explicit
         // if the file doesn't exist it should be a 404
@@ -128,7 +135,7 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
       // /foo -> /fs-root/foo
       if (asSrc && id.startsWith('/')) {
         const fsPath = path.resolve(root, id.slice(1))
-        if ((res = tryFsResolve(fsPath, options, preserveSymlinks))) {
+        if ((res = tryFsResolve(fsPath, options))) {
           isDebug && debug(`[url] ${chalk.cyan(id)} -> ${chalk.dim(res)}`)
           return res
         }
@@ -163,18 +170,12 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
 
         if (
           targetWeb &&
-          (res = tryResolveBrowserMapping(
-            fsPath,
-            importer,
-            options,
-            true,
-            preserveSymlinks
-          ))
+          (res = tryResolveBrowserMapping(fsPath, importer, options, true))
         ) {
           return res
         }
 
-        if ((res = tryFsResolve(fsPath, options, preserveSymlinks))) {
+        if ((res = tryFsResolve(fsPath, options))) {
           isDebug && debug(`[relative] ${chalk.cyan(id)} -> ${chalk.dim(res)}`)
           const pkg = importer != null && idToPkgMap.get(importer)
           if (pkg) {
@@ -189,10 +190,7 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
       }
 
       // absolute fs paths
-      if (
-        path.isAbsolute(id) &&
-        (res = tryFsResolve(id, options, preserveSymlinks))
-      ) {
+      if (path.isAbsolute(id) && (res = tryFsResolve(id, options))) {
         isDebug && debug(`[fs] ${chalk.cyan(id)} -> ${chalk.dim(res)}`)
         return res
       }
@@ -224,13 +222,7 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
 
         if (
           targetWeb &&
-          (res = tryResolveBrowserMapping(
-            id,
-            importer,
-            options,
-            false,
-            preserveSymlinks
-          ))
+          (res = tryResolveBrowserMapping(id, importer, options, false))
         ) {
           return res
         }
@@ -297,7 +289,6 @@ export function resolvePlugin(baseOptions: InternalResolveOptions): Plugin {
 function tryFsResolve(
   fsPath: string,
   options: InternalResolveOptions,
-  preserveSymlinks: boolean,
   tryIndex = true,
   targetWeb = true
 ): string | undefined {
@@ -314,6 +305,23 @@ function tryFsResolve(
   }
 
   let res: string | undefined
+
+  // if we fould postfix exist, we should first try resolving file with postfix. details see #4703.
+  if (
+    postfix &&
+    (res = tryResolveFile(
+      fsPath,
+      '',
+      options,
+      false,
+      targetWeb,
+      options.tryPrefix,
+      options.skipPackageJson
+    ))
+  ) {
+    return res
+  }
+
   if (
     (res = tryResolveFile(
       file,
@@ -321,7 +329,6 @@ function tryFsResolve(
       options,
       false,
       targetWeb,
-      preserveSymlinks,
       options.tryPrefix,
       options.skipPackageJson
     ))
@@ -331,13 +338,27 @@ function tryFsResolve(
 
   for (const ext of options.extensions || DEFAULT_EXTENSIONS) {
     if (
+      postfix &&
+      (res = tryResolveFile(
+        fsPath + ext,
+        '',
+        options,
+        false,
+        targetWeb,
+        options.tryPrefix,
+        options.skipPackageJson
+      ))
+    ) {
+      return res
+    }
+
+    if (
       (res = tryResolveFile(
         file + ext,
         postfix,
         options,
         false,
         targetWeb,
-        preserveSymlinks,
         options.tryPrefix,
         options.skipPackageJson
       ))
@@ -347,13 +368,27 @@ function tryFsResolve(
   }
 
   if (
+    postfix &&
+    (res = tryResolveFile(
+      fsPath,
+      '',
+      options,
+      tryIndex,
+      targetWeb,
+      options.tryPrefix,
+      options.skipPackageJson
+    ))
+  ) {
+    return res
+  }
+
+  if (
     (res = tryResolveFile(
       file,
       postfix,
       options,
       tryIndex,
       targetWeb,
-      preserveSymlinks,
       options.tryPrefix,
       options.skipPackageJson
     ))
@@ -368,51 +403,47 @@ function tryResolveFile(
   options: InternalResolveOptions,
   tryIndex: boolean,
   targetWeb: boolean,
-  preserveSymlinks: boolean,
   tryPrefix?: string,
   skipPackageJson?: boolean
 ): string | undefined {
-  let isReadable = false
-  try {
-    // #2051 if we don't have read permission on a directory, existsSync() still
-    // works and will result in massively slow subsequent checks (which are
-    // unnecessary in the first place)
-    fs.accessSync(file, fs.constants.R_OK)
-    isReadable = true
-  } catch (e) {}
-  if (isReadable) {
+  // #2051 if we don't have read permission on a directory, existsSync() still
+  // works and will result in massively slow subsequent checks (which are
+  // unnecessary in the first place)
+  if (isFileReadable(file)) {
     if (!fs.statSync(file).isDirectory()) {
-      return getRealPath(file, preserveSymlinks) + postfix
+      return getRealPath(file, options.preserveSymlinks) + postfix
     } else if (tryIndex) {
       if (!skipPackageJson) {
         const pkgPath = file + '/package.json'
         if (fs.existsSync(pkgPath)) {
           // path points to a node package
           const pkg = loadPackageData(pkgPath)
-          const resolved = resolvePackageEntry(
-            file,
-            pkg,
-            options,
-            targetWeb,
-            preserveSymlinks
-          )
+          const resolved = resolvePackageEntry(file, pkg, targetWeb, options)
           return resolved
         }
       }
-      const index = tryFsResolve(file + '/index', options, preserveSymlinks)
+      const index = tryFsResolve(file + '/index', options)
       if (index) return index + postfix
     }
   }
-  if (tryPrefix) {
-    const prefixed = `${path.dirname(file)}/${tryPrefix}${path.basename(file)}`
+
+  const tryTsExtension = options.isFromTsImporter && isPossibleTsOutput(file)
+  if (tryTsExtension) {
+    const tsSrcPath = getTsSrcPath(file)
     return tryResolveFile(
-      prefixed,
+      tsSrcPath,
       postfix,
       options,
       tryIndex,
       targetWeb,
-      preserveSymlinks
+      tryPrefix,
+      skipPackageJson
     )
+  }
+
+  if (tryPrefix) {
+    const prefixed = `${path.dirname(file)}/${tryPrefix}${path.basename(file)}`
+    return tryResolveFile(prefixed, postfix, options, tryIndex, targetWeb)
   }
 }
 
@@ -420,7 +451,7 @@ export const idToPkgMap = new Map<string, PackageData>()
 
 export function tryNodeResolve(
   id: string,
-  importer: string | undefined,
+  importer: string | null | undefined,
   options: InternalResolveOptions,
   targetWeb: boolean,
   server?: ViteDevServer,
@@ -435,13 +466,35 @@ export function tryNodeResolve(
   const nestedRoot = id.substring(0, lastArrowIndex).trim()
   const nestedPath = id.substring(lastArrowIndex + 1).trim()
 
-  // check for deep import, e.g. "my-lib/foo"
-  const deepMatch = nestedPath.match(deepImportRE)
+  const possiblePkgIds: string[] = []
+  for (let prevSlashIndex = -1; ; ) {
+    let slashIndex = nestedPath.indexOf('/', prevSlashIndex + 1)
+    if (slashIndex < 0) {
+      slashIndex = nestedPath.length
+    }
 
-  const pkgId = deepMatch ? deepMatch[1] || deepMatch[2] : nestedPath
+    const part = nestedPath.slice(
+      prevSlashIndex + 1,
+      (prevSlashIndex = slashIndex)
+    )
+    if (!part) {
+      break
+    }
+
+    // Assume path parts with an extension are not package roots, except for the
+    // first path part (since periods are sadly allowed in package names).
+    // At the same time, skip the first path part if it begins with "@"
+    // (since "@foo/bar" should be treated as the top-level path).
+    if (possiblePkgIds.length ? path.extname(part) : part[0] === '@') {
+      continue
+    }
+
+    const possiblePkgId = nestedPath.slice(0, slashIndex)
+    possiblePkgIds.push(possiblePkgId)
+  }
 
   let basedir: string
-  if (dedupe && dedupe.includes(pkgId)) {
+  if (dedupe?.some((id) => possiblePkgIds.includes(id))) {
     basedir = root
   } else if (
     importer &&
@@ -453,28 +506,44 @@ export function tryNodeResolve(
     basedir = root
   }
 
-  const preserveSymlinks = !!server?.config.resolve.preserveSymlinks
-
   // nested node module, step-by-step resolve to the basedir of the nestedPath
   if (nestedRoot) {
-    basedir = nestedResolveFrom(nestedRoot, basedir, preserveSymlinks)
+    basedir = nestedResolveFrom(nestedRoot, basedir, options.preserveSymlinks)
   }
 
-  const pkg = resolvePackageData(pkgId, basedir, preserveSymlinks)
+  let pkg: PackageData | undefined
+  const pkgId = possiblePkgIds.reverse().find((pkgId) => {
+    pkg = resolvePackageData(pkgId, basedir, options.preserveSymlinks)
+    return pkg
+  })!
 
   if (!pkg) {
     return
   }
 
-  let resolved = deepMatch
-    ? resolveDeepImport(
-        '.' + id.slice(pkgId.length),
-        pkg,
-        options,
-        targetWeb,
-        preserveSymlinks
-      )
-    : resolvePackageEntry(id, pkg, options, targetWeb, preserveSymlinks)
+  let resolveId = resolvePackageEntry
+  let unresolvedId = pkgId
+  if (unresolvedId !== nestedPath) {
+    resolveId = resolveDeepImport
+    unresolvedId = '.' + nestedPath.slice(pkgId.length)
+  }
+
+  let resolved: string | undefined
+  try {
+    resolved = resolveId(unresolvedId, pkg, targetWeb, options)
+  } catch (err) {
+    if (!options.tryEsmOnly) {
+      throw err
+    }
+  }
+  if (!resolved && options.tryEsmOnly) {
+    resolved = resolveId(unresolvedId, pkg, targetWeb, {
+      ...options,
+      isRequire: false,
+      mainFields: DEFAULT_MAIN_FIELDS,
+      extensions: DEFAULT_EXTENSIONS
+    })
+  }
   if (!resolved) {
     return
   }
@@ -504,7 +573,7 @@ export function tryNodeResolve(
       !isJsType ||
       importer?.includes('node_modules') ||
       exclude?.includes(pkgId) ||
-      exclude?.includes(id) ||
+      exclude?.includes(nestedPath) ||
       SPECIAL_QUERY_RE.test(resolved) ||
       ssr
     ) {
@@ -659,9 +728,8 @@ function loadPackageData(pkgPath: string, cacheKey = pkgPath) {
 export function resolvePackageEntry(
   id: string,
   { dir, data, setResolvedCache, getResolvedCache }: PackageData,
-  options: InternalResolveOptions,
   targetWeb: boolean,
-  preserveSymlinks = false
+  options: InternalResolveOptions
 ): string | undefined {
   const cached = getResolvedCache('.', targetWeb)
   if (cached) {
@@ -698,8 +766,7 @@ export function resolvePackageEntry(
           // instead; Otherwise, assume it's ESM and use it.
           const resolvedBrowserEntry = tryFsResolve(
             path.join(dir, browserEntry),
-            options,
-            preserveSymlinks
+            options
           )
           if (resolvedBrowserEntry) {
             const content = fs.readFileSync(resolvedBrowserEntry, 'utf-8')
@@ -745,11 +812,7 @@ export function resolvePackageEntry(
     }
 
     entryPoint = path.join(dir, entryPoint)
-    const resolvedEntryPoint = tryFsResolve(
-      entryPoint,
-      options,
-      preserveSymlinks
-    )
+    const resolvedEntryPoint = tryFsResolve(entryPoint, options)
 
     if (resolvedEntryPoint) {
       isDebug &&
@@ -805,9 +868,8 @@ function resolveDeepImport(
     dir,
     data
   }: PackageData,
-  options: InternalResolveOptions,
   targetWeb: boolean,
-  preserveSymlinks: boolean
+  options: InternalResolveOptions
 ): string | undefined {
   const cache = getResolvedCache(id, targetWeb)
   if (cache) {
@@ -844,7 +906,6 @@ function resolveDeepImport(
     const resolved = tryFsResolve(
       path.join(dir, relativeId),
       options,
-      preserveSymlinks,
       !exportsField, // try index only if no exports field
       targetWeb
     )
@@ -861,8 +922,7 @@ function tryResolveBrowserMapping(
   id: string,
   importer: string | undefined,
   options: InternalResolveOptions,
-  isFilePath: boolean,
-  preserveSymlinks: boolean
+  isFilePath: boolean
 ) {
   let res: string | undefined
   const pkg = importer && idToPkgMap.get(importer)
@@ -871,7 +931,7 @@ function tryResolveBrowserMapping(
     const browserMappedPath = mapWithBrowserField(mapId, pkg.data.browser)
     if (browserMappedPath) {
       const fsPath = path.join(pkg.dir, browserMappedPath)
-      if ((res = tryFsResolve(fsPath, options, preserveSymlinks))) {
+      if ((res = tryFsResolve(fsPath, options))) {
         isDebug &&
           debug(`[browser mapped] ${chalk.cyan(id)} -> ${chalk.dim(res)}`)
         idToPkgMap.set(res, pkg)