+++ packages/vite/src/node/preview.ts
@@ -3,17 +3,57 @@ import sirv from 'sirv'
 import connect from 'connect'
 import compression from 'compression'
 import { Server } from 'http'
-import { ResolvedConfig, ServerOptions } from '.'
+import { resolveConfig, InlineConfig, ResolvedConfig } from '.'
 import { Connect } from 'types/connect'
+import { ResolvedServerOptions } from './server'
 import {
   resolveHttpsConfig,
   resolveHttpServer,
-  httpServerStart
-} from './server/http'
+  httpServerStart,
+  CommonServerOptions
+} from './http'
 import { openBrowser } from './server/openBrowser'
 import corsMiddleware from 'cors'
 import { proxyMiddleware } from './server/middlewares/proxy'
 import { resolveHostname } from './utils'
+import { printCommonServerUrls } from './logger'
+
+export interface PreviewOptions extends CommonServerOptions {}
+
+export interface ResolvedPreviewOptions extends PreviewOptions {}
+
+export function resolvePreviewOptions(
+  preview: PreviewOptions | undefined,
+  server: ResolvedServerOptions
+): ResolvedPreviewOptions {
+  // The preview server inherits every CommonServerOption from the `server` config
+  // except for the port to enable having both the dev and preview servers running
+  // at the same time without extra configuration
+  return {
+    port: preview?.port,
+    strictPort: preview?.strictPort ?? server.strictPort,
+    host: preview?.host ?? server.host,
+    https: preview?.https ?? server.https,
+    open: preview?.open ?? server.open,
+    proxy: preview?.proxy ?? server.proxy,
+    cors: preview?.cors ?? server.cors
+  }
+}
+
+export interface PreviewServer {
+  /**
+   * The resolved vite config object
+   */
+  config: ResolvedConfig
+  /**
+   * native Node http server instance
+   */
+  httpServer: Server
+  /**
+   * Print server urls
+   */
+  printUrls: () => void
+}
 
 /**
  * Starts the Vite server in preview mode, to simulate a production deployment
@@ -22,24 +62,25 @@ import { resolveHostname } from './utils'
  * @experimental
  */
 export async function preview(
-  config: ResolvedConfig,
-  serverOptions: Pick<ServerOptions, 'port' | 'host'>
-): Promise<Server> {
+  inlineConfig: InlineConfig
+): Promise<PreviewServer> {
+  const config = await resolveConfig(inlineConfig, 'serve', 'production')
+
   const app = connect() as Connect.Server
   const httpServer = await resolveHttpServer(
-    config.server,
+    config.preview,
     app,
-    await resolveHttpsConfig(config)
+    await resolveHttpsConfig(config.preview?.https, config.cacheDir)
   )
 
   // cors
-  const { cors } = config.server
+  const { cors } = config.preview
   if (cors !== false) {
     app.use(corsMiddleware(typeof cors === 'boolean' ? {} : cors))
   }
 
   // proxy
-  if (config.server.proxy) {
+  if (config.preview.proxy) {
     app.use(proxyMiddleware(httpServer, config))
   }
 
@@ -55,9 +96,9 @@ export async function preview(
     })
   )
 
-  const options = config.server
-  const hostname = resolveHostname(serverOptions.host ?? options.host)
-  const port = serverOptions.port ?? 5000
+  const options = config.preview
+  const hostname = resolveHostname(options.host)
+  const port = options.port ?? 5000
   const protocol = options.https ? 'https' : 'http'
   const logger = config.logger
   const base = config.base
@@ -80,5 +121,11 @@ export async function preview(
     )
   }
 
-  return httpServer
+  return {
+    config,
+    httpServer,
+    printUrls() {
+      printCommonServerUrls(httpServer, config.preview, config)
+    }
+  }
 }